got rid of core_types file and moved Orientation/Reflection into response models file. Changed these so they rely on the existing PositionModel class, instead of creating a new Position object
