+++ src/diffcalc_api/__init__.py
@@ -1,6 +1,6 @@
 """API to expose diffcalc-core methods."""
 
-from . import config, core_types, database, openapi, server
+from . import config, database, openapi, server
 from ._version_git import __version__
 
-__all__ = ["__version__", "server", "config", "database", "core_types", "openapi"]
+__all__ = ["__version__", "server", "config", "database", "openapi"]
